+++ changelog.md
@@ -3,6 +3,7 @@
 ## 2.4.0
 
 - Fixed conflict handling for debug keys on 1.21.11.
+- Added a command to switch profiles.
 - Added an optional list entry name field for macros.
 - Changed default Conflict Strategy from SUBMIT to ASSERT.
 - Increased command/message field length limit from 512 to 4096.